Fix test compilation errors
