+++ tests/CacheTower.Tests/CacheStackContextTests.cs
@@ -2,63 +2,28 @@
 using System;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Extensions.Logging.Abstractions;
 
 namespace CacheTower.Tests
 {
 	[TestClass]
 	public class CacheStackContextTests : TestBase
 	{
-		internal class FuncCacheContextActivator<TContext> : ICacheContextActivator
-		{
-			private readonly Func<TContext> Resolver;
-
-			public FuncCacheContextActivator(Func<TContext> resolver)
-			{
-				Resolver = resolver;
-			}
-
-			public ICacheContextScope BeginScope()
-			{
-				return new FuncCacheContextScope<TContext>(Resolver);
-			}
-		}
-
-		internal class FuncCacheContextScope<TContext> : ICacheContextScope
-		{
-			private readonly Func<TContext> Resolver;
-
-			public FuncCacheContextScope(Func<TContext> resolver)
-			{
-				Resolver = resolver;
-			}
-
-			public object Resolve(Type type)
-			{
-				return Resolver()!;
-			}
-
-			public void Dispose()
-			{
-			}
-		}
-
 		[TestMethod, ExpectedException(typeof(ArgumentNullException))]
 		public async Task GetOrSet_ThrowsOnNullKey()
 		{
-			await using var cacheStack = new CacheStack<object>(NullLogger<CacheStack>.Instance, new FuncCacheContextActivator<object>(() => null), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
+			await using var cacheStack = new CacheStack<object>(null, new FuncCacheContextActivator<object>(() => null), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
 			await cacheStack.GetOrSetAsync<int>(null, (old, context) => Task.FromResult(5), new CacheSettings(TimeSpan.FromDays(1)));
 		}
 		[TestMethod, ExpectedException(typeof(ArgumentNullException))]
 		public async Task GetOrSet_ThrowsOnNullGetter()
 		{
-			await using var cacheStack = new CacheStack<object>(NullLogger<CacheStack>.Instance, new FuncCacheContextActivator<object>(() => null), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
+			await using var cacheStack = new CacheStack<object>(null, new FuncCacheContextActivator<object>(() => null), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
 			await cacheStack.GetOrSetAsync<int>("MyCacheKey", null, new CacheSettings(TimeSpan.FromDays(1)));
 		}
 		[TestMethod]
 		public async Task GetOrSet_CacheMiss_ContextHasValue()
 		{
-			await using var cacheStack = new CacheStack<int>(NullLogger<CacheStack>.Instance, new FuncCacheContextActivator<object>(() => 123), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
+			await using var cacheStack = new CacheStack<int>(null, new FuncCacheContextActivator<object>(() => 123), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
    			var result = await cacheStack.GetOrSetAsync<int>("GetOrSet_CacheMiss_ContextHasValue", (oldValue, context) =>
 			{
 				Assert.AreEqual(123, context);
@@ -71,7 +36,7 @@ namespace CacheTower.Tests
 		public async Task GetOrSet_CacheMiss_ContextFactoryCalledEachTime()
 		{
 			var contextValue = 0;
-			await using var cacheStack = new CacheStack<int>(NullLogger<CacheStack>.Instance, new FuncCacheContextActivator<object>(() => contextValue++), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
+			await using var cacheStack = new CacheStack<int>(null, new FuncCacheContextActivator<object>(() => contextValue++), new[] { new MemoryCacheLayer() }, Array.Empty<ICacheExtension>());
 
 			var result1 = await cacheStack.GetOrSetAsync<int>("GetOrSet_CacheMiss_ContextFactoryCalledEachTime_1", (oldValue, context) =>
 			{
